Moved get_neighbouring_tiles to TileBatch
