+++ src/collisions/interface/Collisions.hpp
@@ -1,9 +1,6 @@
 #pragma once
 
-#include <cstdint>
-
 #include "MapTile.hpp"
-#include "TileBatch.hpp"
 
 namespace collisions
 {
@@ -21,23 +18,7 @@ namespace collisions
     *
     * In example, tile number 4 is the CENTER tile in this 9-element neighboring tiles array.
     */
-    enum class NeighbouringTiles : std::uint16_t
-    {
-        LEFT_MIDDLE = 0,
-        RIGHT_MIDDLE = 1,
-        UP_MIDDLE = 2,
-        DOWN_MIDDLE = 3,
-        CENTER = 4,
-        LEFT_UP = 5,
-        RIGHT_UP = 6,
-        LEFT_DOWN = 7,
-        RIGHT_DOWN = 8,
-    };
-
     bool overlaps(MapTile* tile, float x_center, float y_center, float width, float height);
     MapTile* overlaps(MapTile* neighboring_tiles[9], float x_center, float y_center, float width, float height, bool collidable = true);
     MapTile* overlaps_strict(MapTile *neighboring_tiles[9], float x_center, float y_center, float width, float height, bool collidable = true);
-
-    // FIXME: Probably should be a method of the Level/TileBatch class.
-    void get_neighbouring_tiles(const TileBatch&, float x, float y, MapTile *out_neighboring_tiles[9]);
 }
