+++ src/level/src/TileBatch.cpp
@@ -8,6 +8,7 @@
 #include <cstring>
 #include <algorithm>
 
+#include "NeighbouringTiles.hpp"
 #include "Level.hpp"
 #include "Renderer.hpp"
 #include "TileBatch.hpp"
@@ -548,4 +549,46 @@ void TileBatch::generate_cave_background()
             }
         }
     }
+}
+
+void TileBatch::get_neighbouring_tiles(float x, float y, MapTile *out_neighboring_tiles[9]) const
+{
+    const float width = 1.0f;
+    const float height = 1.0f;
+
+    std::uint16_t x_tiles = std::floor(x + (width / 2));
+    std::uint16_t y_tiles = std::floor(y + (height / 2));
+
+    assert(x_tiles < Consts::LEVEL_WIDTH_TILES);
+    assert(y_tiles < Consts::LEVEL_HEIGHT_TILES);
+
+    MapTile *left_middle = nullptr,
+            *right_middle = nullptr,
+            *up_middle = nullptr,
+            *down_middle = nullptr,
+            *center = nullptr,
+            *left_up = nullptr,
+            *right_up = nullptr,
+            *left_down = nullptr,
+            *right_down = nullptr;
+
+    left_middle = x_tiles - 1 >= 0 && map_tiles[x_tiles - 1][y_tiles]->exists ? map_tiles[x_tiles - 1][y_tiles] : nullptr;
+    right_middle = x_tiles + 1 <= 31 && map_tiles[x_tiles + 1][y_tiles]->exists ? map_tiles[x_tiles + 1][y_tiles] : nullptr;
+    up_middle = y_tiles - 1 >= 0 && map_tiles[x_tiles][y_tiles - 1]->exists ? map_tiles[x_tiles][y_tiles - 1] : nullptr;
+    down_middle = y_tiles + 1 <= 31 && map_tiles[x_tiles][y_tiles + 1]->exists ? map_tiles[x_tiles][y_tiles + 1] : nullptr;
+    center = map_tiles[x_tiles][y_tiles]->exists ? map_tiles[x_tiles][y_tiles] : nullptr;
+    left_up = x_tiles - 1 >= 0 && y_tiles - 1 >= 0 && map_tiles[x_tiles - 1][y_tiles - 1]->exists ? map_tiles[x_tiles - 1][y_tiles - 1] : nullptr;
+    right_up = x_tiles + 1 <= 31 && y_tiles - 1 >= 0 && map_tiles[x_tiles + 1][y_tiles - 1]->exists ? map_tiles[x_tiles + 1][y_tiles - 1] : nullptr;
+    left_down = x_tiles - 1 >= 0 && y_tiles + 1 <= 31 && map_tiles[x_tiles - 1][y_tiles + 1]->exists ? map_tiles[x_tiles - 1][y_tiles + 1] : nullptr;
+    right_down = x_tiles + 1 <= 31 && y_tiles + 1 <= 31 && map_tiles[x_tiles + 1][y_tiles + 1]->exists ? map_tiles[x_tiles + 1][y_tiles + 1] : nullptr;
+
+    out_neighboring_tiles[static_cast<uint16_t>(NeighbouringTiles::LEFT_MIDDLE)] = left_middle;
+    out_neighboring_tiles[static_cast<uint16_t>(NeighbouringTiles::RIGHT_MIDDLE)] = right_middle;
+    out_neighboring_tiles[static_cast<uint16_t>(NeighbouringTiles::UP_MIDDLE)] = up_middle;
+    out_neighboring_tiles[static_cast<uint16_t>(NeighbouringTiles::DOWN_MIDDLE)] = down_middle;
+    out_neighboring_tiles[static_cast<uint16_t>(NeighbouringTiles::CENTER)] = center;
+    out_neighboring_tiles[static_cast<uint16_t>(NeighbouringTiles::LEFT_UP)] = left_up;
+    out_neighboring_tiles[static_cast<uint16_t>(NeighbouringTiles::RIGHT_UP)] = right_up;
+    out_neighboring_tiles[static_cast<uint16_t>(NeighbouringTiles::LEFT_DOWN)] = left_down;
+    out_neighboring_tiles[static_cast<uint16_t>(NeighbouringTiles::RIGHT_DOWN)] = right_down;
 }
